% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abridged2single_1014_df.R
\docType{data}
\name{abridged2single_1014_df}
\alias{abridged2single_1014_df}
\title{abridged2single_1014_df}
\format{
A data frame with 38 rows and 12 variables:
\describe{
...
}
}
\usage{
abridged2single_1014_df
}
\description{
This dataset has been used to show how dd_births_abridged2single_1014() works.
}
\keyword{internal}
