% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_locid.R
\name{check_locid}
\alias{check_locid}
\title{check_locid}
\usage{
check_locid(locid)
}
\arguments{
\item{locid}{location id}
}
\value{
The message: \code{locid} is not a valid location id. Please run View(get_locations()) to get a list of plausible location ids. They are listed in the \code{PK_LocID} variable
}
\description{
Produces an error message if a location id is not part of the possible location ids in the UNDP database
}
\examples{
check_locid(12345)
}
