% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_abr.R
\docType{data}
\name{data_abr}
\alias{data_abr}
\title{data_abr}
\format{
A data frame with 11 rows and 12 variables:
\describe{
...
}
}
\usage{
data_abr
}
\description{
This dataset has been used to show how DDharmonize_AbridgedAndComplete() works.
}
\keyword{internal}
