% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cpl.R
\docType{data}
\name{data_cpl}
\alias{data_cpl}
\title{data_cpl}
\format{
A data frame with 11 rows and 12 variables:
\describe{
...
}
}
\usage{
data_cpl
}
\description{
This dataset has been used to show how DDharmonize_AbridgedAndComplete() works.
}
\keyword{datasets}
