% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_age_standard.R
\name{dd_age_standard}
\alias{dd_age_standard}
\title{dd_age_standard}
\usage{
dd_age_standard(data, abridged = TRUE)
}
\arguments{
\item{data}{The data to be harmonized}

\item{abridged}{\code{TRUE/FALSE}}
}
\value{
The DemoData dataset with standard age groups appended to it.
}
\description{
Add standard age groups in the dataset
}
\examples{
df <- firstages_compute_births_df
df <- dd_age_standard(df)
}
