% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_distribute_unknowns.R
\name{dd_distribute_unknowns}
\alias{dd_distribute_unknowns}
\title{dd_distribute_unknowns}
\usage{
dd_distribute_unknowns(data)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
Distribute Unknowns over age
}
