% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_fillzeros_births.R
\name{dd_fillzeros_births}
\alias{dd_fillzeros_births}
\title{dd_fillzeros_births}
\usage{
dd_fillzeros_births(data, abridged = TRUE)
}
\arguments{
\item{data}{The data to be harmonized}

\item{abridged}{\code{TRUE/FALSE}}
}
\value{

}
\description{
Fill in zeros for mother ages younger than 15 if there are no data reported
}
