% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_firstages_compute.R
\name{dd_firstages_compute}
\alias{dd_firstages_compute}
\title{dd_firstages_compute
Reconcile ages 0-1, 1-4 and 0-4}
\usage{
dd_firstages_compute(data)
}
\arguments{
\item{data}{The data to be harmonized}
}
\value{
A dataset with ages 0-1, 1-4 and 0-4 reconciled
}
\description{
dd_firstages_compute
Reconcile ages 0-1, 1-4 and 0-4
}
