% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_firstages_compute_births.R
\name{dd_firstages_compute_births}
\alias{dd_firstages_compute_births}
\title{dd_firstages_compute_births}
\usage{
dd_firstages_compute_births(data)
}
\arguments{
\item{data}{The data to be harmonized}
}
\value{
A dataset with early age groups reconciled e.g 0-14 becomes 10-14
}
\description{
Reconcile early age groups for births
}
\examples{
df <- vitals5_wags
df <- dd_firstages_compute_births(vitals5_wags)
}
