% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_latest_source_year.R
\name{dd_latest_source_year}
\alias{dd_latest_source_year}
\title{dd_latest_source_year
Select the latest data source year. For some periods of reference, there are multiple values in the \strong{DataSourceYear} variable. For these cases, the criteria is to choose the latest \strong{DataSourceYear}}
\usage{
dd_latest_source_year(data)
}
\arguments{
\item{data}{The data to be harmonized}
}
\value{
A dataset with a unique \strong{DataSourceYear}
}
\description{
dd_latest_source_year
Select the latest data source year. For some periods of reference, there are multiple values in the \strong{DataSourceYear} variable. For these cases, the criteria is to choose the latest \strong{DataSourceYear}
}
