% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_oag_agestart.R
\name{dd_oag_agestart}
\alias{dd_oag_agestart}
\title{dd_oag_agestart}
\usage{
dd_oag_agestart(data, multiple5 = TRUE)
}
\arguments{
\item{data}{The dataset to be harmonized}

\item{multiple5}{Whether the series contains 5 year age labels or not}
}
\value{
An age label that would close the series. It has to be a multiple of 5
}
\description{
Identifies the starting age needed for the open age group that closes the series. It must be a multiple of 5
}
\details{
The function returns the starting age needed for the open age group that closes the series. This corresponds to the
\code{AgeEnd} of the maximum \code{AgeStart} in the series, for records with an \code{AgeSpan > 0} per id, sex id and time label.
You can run \code{getAnywhere(dd_oag_agestart)} to see the function definition.
}
\examples{
\dontrun{
df <- series_isfull_df \%>\%
select(IndicatorID, IndicatorName, id, SexName, TimeLabel,AgeSpan, AgeStart, AgeEnd, AgeLabel, DataValue) \%>\%
arrange(AgeLabel)
oag_start <- dd_oag_agestart(df, multiple5 = TRUE)
}
}
