% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_oag_compute.R
\name{dd_oag_compute}
\alias{dd_oag_compute}
\title{dd_oag_compute}
\usage{
dd_oag_compute(data, age_span = c(1, 5))
}
\arguments{
\item{data}{The data to be harmonized}

\item{age_span}{xxx}
}
\value{
A dataset with plausible closing age groups and their data values
}
\description{
Compute data values for the open age group needed to close the series
}
