% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_oag_multiple.R
\name{dd_oag_multiple}
\alias{dd_oag_multiple}
\title{dd_oag_multiple}
\usage{
dd_oag_multiple(data)
}
\arguments{
\item{data}{The data to be harmonised}
}
\value{
\code{TRUE/FALSE}
}
\description{
Check whether the series has more than one open age group
}
