% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_series_isfull.R
\name{dd_series_isfull}
\alias{dd_series_isfull}
\title{dd_series_isfull}
\usage{
dd_series_isfull(data, abridged = TRUE)
}
\arguments{
\item{data}{The dataset to be harmonized}

\item{abridged}{Whether the data is abridged or not. If the data contains 5-year age labels, \code{abridged == TRUE}}
}
\value{
\code{TRUE/FALSE}. TRUE if the series is full and FALSE otherwise
}
\description{
Check whether the age vector meets conditions for a full series.
A full series requires all closed age groups and either an open age group or a total. If abridged, then a full series also requires either both abridged child ages or both five-year child ages to be present
}
\examples{
df <- series_isfull_df
dd_series_isfull(df, abridged = TRUE)
}
