% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_single2abridged.R
\name{dd_single2abridged}
\alias{dd_single2abridged}
\title{dd_single2abridged}
\usage{
dd_single2abridged(data)
}
\arguments{
\item{data}{dataset with age series by single years of age}
}
\value{
dataset containing 5-year age groups generated from single years of age
}
\description{
Creates 5-year age groups based on data by single years of age.
}
