% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_validate_totals_over_age.R
\name{dd_validate_totals_over_age}
\alias{dd_validate_totals_over_age}
\title{dd_validate_totals_over_age}
\usage{
dd_validate_totals_over_age(data)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
Check if Totals are equal or different from previous values, after having standardized age series
}
