% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_validate_totals_over_sex.R
\name{dd_validate_totals_over_sex}
\alias{dd_validate_totals_over_sex}
\title{dd_validate_totals_over_sex}
\usage{
dd_validate_totals_over_sex(data)
}
\arguments{
\item{data}{Data to be harmonized}
}
\value{
A dataset where Both sexes == Female + Male
}
\description{
Check if Population counts by sex is equal to population counts for both sexes and adjust in case of any difference.
Criteria: If Female Population + Male Population is different from the population for Both sexes, then Both sexes = Female + Male
}
\details{
See the \href{https://shelmith-kariuki.github.io/rddharmony/articles/dd_validate_totals_over_sex.html}{"Validating totals over sex" vignette} for more details about this function.
}
