% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_validate_totals_over_sex_new.R
\name{dd_validate_totals_over_sex_new}
\alias{dd_validate_totals_over_sex_new}
\title{dd_validate_totals_over_sex_new}
\usage{
dd_validate_totals_over_sex_new(data)
}
\arguments{
\item{data}{Data to be harmonized}
}
\value{
A dataset where Both sexes == Female + Male
}
\description{
Compute total over sex as max of reported total or sum over sex, compute any NA values based on differences and
redistribute SexID = 0 across males and females
}
