% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDextract_VitalCounts.R
\name{DDextract_VitalCounts}
\alias{DDextract_VitalCounts}
\title{DDextract_VitalCounts}
\usage{
DDextract_VitalCounts(
  locid,
  type = c("births", "deaths"),
  process = c("census", "vr"),
  start_year,
  end_year,
  DataSourceShortName = NULL,
  DataSourceYear = NULL
)
}
\arguments{
\item{locid}{The Location Id for each country. Run \code{DemoTools::get_locations()} to get information about available locations. The ids are indicated in the \code{PK_LocID} variable.}

\item{type}{The type of data to be pulled i.e. births/deaths}

\item{process}{The data collection process i.e. Census (census) or Vital Registrations (vr)}

\item{start_year}{The minimum year for the data available for each country}

\item{end_year}{The maximum year for the data available for each country}

\item{DataSourceShortName}{NULL.}

\item{DataSourceYear}{NULL.}
}
\value{
A dataset showing counts for each location, type of data (births or deaths), process, year, sex and age label.
}
\description{
Extracts births and/or deaths data for a given country and time period from the UNDP portal.
}
\details{
Testing that this section shows up at the Details section of the help file
}
\examples{
dd_extract <- DDextract_VitalCounts(404, #Kenya
                                   type = c("births"),
                                   process = c("census","vr"),
                                   1950,
                                   2050,
                                   DataSourceShortName = NULL,
                                   DataSourceYear = NULL)
}
