% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDharmonize_AbridgedAndComplete.R
\name{DDharmonize_AbridgedAndComplete}
\alias{DDharmonize_AbridgedAndComplete}
\title{DDharmonize_AbridgedAndComplete}
\usage{
DDharmonize_AbridgedAndComplete(data_abr, data_cpl_from_abr, data_cpl)
}
\arguments{
\item{data_abr}{data containing abridged series age labels}

\item{data_cpl_from_abr}{data containing complete series derived from abridged series}

\item{data_cpl}{data containing complete series age labels}
}
\value{
A dataset with reconciled abridged and complete series i.e. abridged reconciled with complete and complete reconciled with abridged
}
\description{
This wrapper takes the output of harmonization and reconciles abridged and complete series
}
