% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDharmonize_Pop1.R
\name{DDharmonize_Pop1}
\alias{DDharmonize_Pop1}
\title{DDharmonize_Pop1}
\usage{
DDharmonize_Pop1(indata)
}
\arguments{
\item{indata}{The data to be harmonised}
}
\value{
A data frame that contains data by single year of age up to the open age group
}
\description{
This wrapper takes population1 indicators from DemoData and Standardizes/harmonizes by age group
}
\examples{
df <- data.frame()
}
