% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDharmonize_validate_BirthCounts.R
\name{DDharmonize_validate_BirthCounts}
\alias{DDharmonize_validate_BirthCounts}
\title{DDharmonize_validate_BirthCounts}
\usage{
DDharmonize_validate_BirthCounts(
  locid,
  times,
  process = c("census", "vr"),
  return_unique_ref_period = TRUE,
  DataSourceShortName = NULL,
  DataSourceYear = NULL,
  retainKeys = FALSE,
  server = "https://popdiv.dfs.un.org/DemoData/api/"
)
}
\arguments{
\item{locid}{Location id}

\item{times}{1950, 2050}

\item{process}{census or vr}

\item{return_unique_ref_period}{TRUE}

\item{DataSourceShortName}{NULL}

\item{DataSourceYear}{NULL}

\item{retainKeys}{FALSE}

\item{server}{"https://popdiv.dfs.un.org/DemoData/api/"}
}
\value{
A harmonized dataset
}
\description{
THIS SCRIPT IMPLEMENTS A WORKFLOW FOR Births COUNTS from vr or census EXTRACTING FROM DEMODATA, HARMONIZING AGE GROUPS, IDENTIFYING FULL SERIES, SELECTING PREFERRED SERIES, VALIDATING TOTALS AND BY SEX AND, EVENTUALLY, use retainKeys = TRUE to include these key fields in the function output.
}
\examples{
sweden_df <- DDharmonize_validate_BirthCounts(752,
                                              c(1950,2020),
                                              process = c("census", "vr"),
                                              return_unique_ref_period = TRUE, # if true, then only most authoratative series will be returned for each reference period, per dd_rank_id_vitals()
                                              DataSourceShortName = NULL,
                                              DataSourceYear = NULL,
                                              retainKeys = FALSE,
                                              server = "https://popdiv.dfs.un.org/DemoData/api/")
}
