% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDharmonize_validate_DeathCounts.R
\name{DDharmonize_validate_DeathCounts}
\alias{DDharmonize_validate_DeathCounts}
\title{DDharmonize_validate_DeathCounts}
\usage{
DDharmonize_validate_DeathCounts(
  locid,
  times,
  process = c("census", "vr"),
  return_unique_ref_period = TRUE,
  DataSourceShortName = NULL,
  DataSourceYear = NULL,
  retainKeys = FALSE,
  server = "https://popdiv.dfs.un.org/DemoData/api/"
)
}
\arguments{
\item{locid}{location id}

\item{times}{1950, 2050}

\item{process}{census or vr}

\item{return_unique_ref_period}{TRUE}

\item{DataSourceShortName}{NULL}

\item{DataSourceYear}{NULL}

\item{retainKeys}{FALSE}

\item{server}{"https://popdiv.dfs.un.org/DemoData/api/"}
}
\value{
A harmonized dataset containing death counts
}
\description{
This script implements a workflow for death counts from vr or census extracting data from Demodata, harmonizing age groups,
identifying full series, selecting preferred series, validating totals by age and by sex and eventually,
including key fields in the function output.
}
\examples{
kenya_df <- DDharmonize_validate_DeathCounts(404,
                                              c(1950,2020),
                                              process = c("census", "vr"),
                                              return_unique_ref_period = TRUE,
                                              DataSourceShortName = NULL,
                                              DataSourceYear = NULL,
                                              retainKeys = FALSE,
                                              server = "https://popdiv.dfs.un.org/DemoData/api/")
}
