% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dups.R
\docType{data}
\name{dups}
\alias{dups}
\title{dups}
\format{
A data frame with one variable that shows the duplicates to be discarded.
\describe{
...
}
}
\usage{
dups
}
\description{
This dataset contains duplicates that should be dropped from the data. It has been used in the \code{dd_rank_id_vitals()} function
}
\keyword{internal}
