% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstages_compute_df.R
\docType{data}
\name{firstages_compute_df}
\alias{firstages_compute_df}
\title{firstages_compute_df}
\format{
A data frame with 11 rows and 6 variables:
\describe{
...
}
}
\usage{
firstages_compute_df
}
\description{
This dataset has been used to show how dd_firstages_compute() works.
}
\keyword{internal}
