% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latest_source_year_df.R
\docType{data}
\name{latest_source_year_df}
\alias{latest_source_year_df}
\title{latest_source_year_df}
\format{
A data frame with 12 rows and 95 variables:
\describe{
...
}
}
\usage{
latest_source_year_df
}
\description{
This dataset has been used to show how dd_latest_source_year() works.
}
\keyword{datasets}
