% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oag_compute_df.R
\docType{data}
\name{oag_compute_df}
\alias{oag_compute_df}
\title{oag_compute_df}
\format{
A data frame with 11 rows and 6 variables:
\describe{
...
}
}
\usage{
oag_compute_df
}
\description{
This dataset has been used to show how dd_oag_compute_df() works.
}
\keyword{datasets}
