% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oag_multiple_df.R
\docType{data}
\name{oag_multiple_df}
\alias{oag_multiple_df}
\title{oag_multiple_df}
\format{
A data frame with 11 rows and 7 variables:
\describe{
...
}
}
\usage{
oag_multiple_df
}
\description{
This dataset has been used to show how dd_oag_multiple() works.
}
\keyword{datasets}
