% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop1_df.R
\docType{data}
\name{pop1_df}
\alias{pop1_df}
\title{pop1_df}
\format{
A data frame with 97 rows and 95 variables:
\describe{
...
}
}
\usage{
pop1_df
}
\description{
This dataset has been used to show how DDharmonize_Pop1() works
}
\keyword{datasets}
