% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_dropped_ids.R
\name{print_dropped_ids}
\alias{print_dropped_ids}
\title{Title}
\usage{
print_dropped_ids(indata_before, indata_after)
}
\arguments{
\item{indata_before}{vitals_std_all}

\item{indata_after}{vitals_std_full}
}
\value{

}
\description{
Title
}
