% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_isfull_df.R
\docType{data}
\name{series_isfull_df}
\alias{series_isfull_df}
\title{series_isfull_df}
\format{
A data frame with 10 rows and 94 variables:
\describe{
...
}
}
\usage{
series_isfull_df
}
\description{
This dataset has been used to show how dd_oag_agestart() and dd_series_isfull() work.
}
\keyword{datasets}
