% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_totals_over_sex_bd.R
\docType{data}
\name{validate_totals_over_sex_bd}
\alias{validate_totals_over_sex_bd}
\title{validate_totals_over_sex_bd}
\format{
A data frame with 360 rows and 48 variables.

\describe{
...
}
}
\usage{
validate_totals_over_sex_bd
}
\description{
This dataset has been used to show how dd_validate_totals_over_sex() works.
}
\keyword{datasets}
