% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_totals_over_sex_pop.R
\docType{data}
\name{validate_totals_over_sex_pop}
\alias{validate_totals_over_sex_pop}
\title{validate_totals_over_sex_pop}
\format{
A data frame with 63 rows and 46 variables.

\describe{
...
}
}
\usage{
validate_totals_over_sex_pop
}
\description{
This dataset has been used to show how dd_validate_totals_over_sex_new() works.
}
\keyword{internal}
