% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vitals1_std.R
\docType{data}
\name{vitals1_std}
\alias{vitals1_std}
\title{vitals1_std}
\format{
A data frame with 38 rows and 12 variables:
\describe{
...
}
}
\usage{
vitals1_std
}
\description{
This dataset has been used to show how dd_births_abridged2single_1014() works.
}
\keyword{datasets}
