% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vitals5_df.R
\docType{data}
\name{vitals5_df}
\alias{vitals5_df}
\title{vitals5_df}
\format{
A data frame with 11 rows and 95 variables:
\describe{
...
}
}
\usage{
vitals5_df
}
\description{
This dataset has been used to show how DDharmonize_Vitals5() works
}
\keyword{datasets}
