% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vitals5_std.R
\docType{data}
\name{vitals5_std}
\alias{vitals5_std}
\title{vitals5_std}
\format{
A data frame with 11 rows and 12 variables:
\describe{
...
}
}
\usage{
vitals5_std
}
\description{
vitals5_std
}
\keyword{internal}
