% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vitals_abr_cpl1_f0.R
\docType{data}
\name{vitals_abr_cpl1_f0}
\alias{vitals_abr_cpl1_f0}
\title{vitals_abr_cpl1_f0}
\format{
A data frame with 11 rows and 11 variables:
\describe{
...
}
}
\usage{
vitals_abr_cpl1_f0
}
\description{
This dataset has been used to show how dd_fillzeros_births() works, in a case where abridged = TRUE
}
\keyword{internal}
