% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vitals_abr_cpl2.R
\docType{data}
\name{vitals_abr_cpl2}
\alias{vitals_abr_cpl2}
\title{vitals_abr_cpl2}
\format{
A data frame with 43 rows and 12 variables:
\describe{
...
}
}
\usage{
vitals_abr_cpl2
}
\description{
This dataset has been used to show how dd_single2abridged() works.
}
\keyword{internal}
