% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vitals_std_full_sample.R
\docType{data}
\name{vitals_std_full_sample}
\alias{vitals_std_full_sample}
\title{vitals_std_full_sample}
\format{
A data frame with 68 rows and 48 variables:
\describe{
...
}
}
\usage{
vitals_std_full_sample
}
\description{
This dataset has been used to show how dd_validate_totals_over_age() works.
}
\keyword{datasets}
