% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vitals_std_valid_sample2.R
\docType{data}
\name{vitals_std_valid_sample2}
\alias{vitals_std_valid_sample2}
\title{vitals_std_valid_sample2}
\format{
A data frame with 2510 rows and 49 variables:
\describe{
...
}
}
\usage{
vitals_std_valid_sample2
}
\description{
This dataset has been used to show how dd_rank_id_vitals() works.
}
\keyword{internal}
