% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_firstages_compute_births.R
\name{dd_firstages_compute_births}
\alias{dd_firstages_compute_births}
\title{dd_firstages_compute_births}
\usage{
dd_firstages_compute_births(data)
}
\arguments{
\item{data}{The data to be reconciled}
}
\value{
A dataset with early age groups reconciled
}
\description{
Reconcile early age groups for births
}
