% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_oag_agestart.R
\name{dd_oag_agestart}
\alias{dd_oag_agestart}
\title{dd_oag_agestart}
\usage{
dd_oag_agestart(data, multiple5 = TRUE)
}
\arguments{
\item{data}{The dataset to be harmonized}

\item{multiple5}{Whether the series contains 5 year age labels or not}
}
\value{
A value that has to be a multiple of 5
}
\description{
Identify the starting age needed for the open age group that closes the series. It must be a multiple of 5
}
