% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDharmonize_validate_BirthCounts.R
\name{DDharmonize_validate_BirthCounts}
\alias{DDharmonize_validate_BirthCounts}
\title{DDharmonize_validate_BirthCounts}
\usage{
DDharmonize_validate_BirthCounts(
  locid,
  times,
  process = c("census", "vr"),
  return_unique_ref_period = TRUE,
  DataSourceShortName = NULL,
  DataSourceYear = NULL,
  retainKeys = FALSE,
  server = "https://popdiv.dfs.un.org/DemoData/api/"
)
}
\arguments{
\item{locid}{location id}

\item{times}{1950, 2050}

\item{process}{census or vr}

\item{return_unique_ref_period}{TRUE}

\item{DataSourceShortName}{NULL}

\item{DataSourceYear}{NULL}

\item{retainKeys}{FALSE}

\item{server}{"https://popdiv.dfs.un.org/DemoData/api/"}
}
\value{
A harmonized dataset
}
\description{
This script implements a workflow for births counts from vr or census extracting data from Demodata, harmonizing age groups,
identifying full series, selecting preferred series, validating totals by age and by sex and, eventually,
including key fields in the function output.
}
\examples{
sweden_df <- DDharmonize_validate_BirthCounts(752,
                                              c(1950,2020),
                                              process = c("census", "vr"),
                                              return_unique_ref_period = TRUE,
                                              DataSourceShortName = NULL,
                                              DataSourceYear = NULL,
                                              retainKeys = FALSE,
                                              server = "https://popdiv.dfs.un.org/DemoData/api/")
}
