% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDharmonize_validate_DeathCounts.R
\name{DDharmonize_validate_DeathCounts}
\alias{DDharmonize_validate_DeathCounts}
\title{DDharmonize_validate_DeathCounts}
\usage{
DDharmonize_validate_DeathCounts(
  locid,
  times,
  process = c("census", "vr"),
  return_unique_ref_period = TRUE,
  DataSourceShortName = NULL,
  DataSourceYear = NULL,
  retainKeys = FALSE,
  server = "https://popdiv.dfs.un.org/DemoData/api/"
)
}
\arguments{
\item{locid}{location id}

\item{times}{The period of the data to be extracted. You can extract one year data e.g times = 2020 or a longer period of time e.g times = c(1950, 2020).}

\item{process}{The process through which the data was obtained from various sources i.e either via census or vital registrations (vr). By default, the function pulls data obtained through both of these processes.}

\item{return_unique_ref_period}{TRUE}

\item{DataSourceShortName}{NULL}

\item{DataSourceYear}{NULL}

\item{retainKeys}{FALSE}

\item{server}{"https://popdiv.dfs.un.org/DemoData/api/"}
}
\value{
A harmonized dataset containing death counts
}
\description{
This script implements a workflow for death counts from vr or census extracting data from Demodata, harmonizing age groups,
identifying full series, selecting preferred series, validating totals by age and by sex and eventually,
including key fields in the function output.
}
\examples{
\dontrun{
kenya_df <- DDharmonize_validate_DeathCounts(404,
                                              c(1950,2020),
                                              process = c("census", "vr"),
                                              return_unique_ref_period = TRUE,
                                              DataSourceShortName = NULL,
                                              DataSourceYear = NULL,
                                              retainKeys = FALSE,
                                              server = "https://popdiv.dfs.un.org/DemoData/api/")
}
}
