% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDharmonize_validate_PopCounts.R
\name{DDharmonize_validate_PopCounts}
\alias{DDharmonize_validate_PopCounts}
\title{DDharmonize_validate_PopCounts}
\usage{
DDharmonize_validate_PopCounts(
  locid,
  times,
  process = c("census", "estimate", "register"),
  return_unique_ref_period = TRUE,
  DataSourceShortName = NULL,
  DataSourceYear = NULL,
  retainKeys = FALSE,
  server = "https://popdiv.dfs.un.org/DemoData/api/"
)
}
\arguments{
\item{locid}{location id}

\item{times}{The period of the data to be extracted. You can extract one year data e.g times = 2020 or a longer period of time e.g times = c(1950, 2020).}

\item{process}{The process through which the data was obtained from various sources i.e either via census , register or estimates. By default, the function pulls data obtained through both of these processes.}

\item{return_unique_ref_period}{TRUE}

\item{DataSourceShortName}{NULL}

\item{DataSourceYear}{NULL}

\item{retainKeys}{FALSE}

\item{server}{"https://popdiv.dfs.un.org/DemoData/api/"}
}
\value{
A harmonized dataset containing birth counts
}
\description{
This script implements a work flow for census population counts extracting from DemoData, harmonizing age groups,
identifying full series, selecting preferred series, validating totals and by sex and eventually
including key fields in the function output.
}
\examples{
sweden_df <- DDharmonize_validate_PopCounts(locid = 752,
                                              times = c(1950,2020),
                                              process = c("census", "estimate", "register"),
                                              return_unique_ref_period = TRUE,
                                              DataSourceShortName = NULL,
                                              DataSourceYear = NULL,
                                              retainKeys = FALSE,
                                              server = "https://popdiv.dfs.un.org/DemoData/api/")
}
